requirements.txt
streamlit>=1.35.0
google-generativeai>=0.8.6
beautifulsoup4
requests